refactor gpu_metrics code
