+++ tb_plugin/torch_tb_profiler/run.py
@@ -3,6 +3,8 @@
 # --------------------------------------------------------------------------
 from collections import OrderedDict
 
+from . import io
+
 
 class Run(object):
     """ A profiler run. For visualization purpose only.
@@ -63,6 +65,51 @@ class RunProfile(object):
         self.gpu_util_buckets = None
         self.approximated_sm_efficency_ranges = None
 
+    def get_gpu_metrics(self):
+        def build_trace_counter_gpu_util(gpu_id, start_time, counter_value):
+            util_json = ", {{\"ph\":\"C\", \"name\":\"GPU {} Utilization\", " \
+                        "\"pid\":{}, \"ts\":{}, " \
+                        "\"args\":{{\"GPU Utilization\":{}}}}}".format(
+                gpu_id, gpu_id, start_time, counter_value
+            )
+            return util_json
+
+        def build_trace_counter_sm_efficiency(gpu_id, start_time, counter_value):
+            util_json = ", {{\"ph\":\"C\", \"name\":\"GPU {} Est. SM Efficiency\", " \
+                        "\"pid\":{}, \"ts\":{}, " \
+                        "\"args\":{{\"Est. SM Efficiency\":{}}}}}".format(
+                gpu_id, gpu_id, start_time, counter_value
+            )
+            return util_json
+
+        counter_json_str = ""
+        for gpu_id in range(len(self.gpu_util_buckets)):
+            buckets = self.gpu_util_buckets[gpu_id]
+            for b in buckets:
+                json_str = build_trace_counter_gpu_util(gpu_id, b[0], b[1])
+                counter_json_str += json_str
+        for gpu_id in range(len(self.approximated_sm_efficency_ranges)):
+            ranges = self.approximated_sm_efficency_ranges[gpu_id]
+            for r in ranges:
+                efficiency_json_start = build_trace_counter_sm_efficiency(gpu_id, r[0][0], r[1])
+                efficiency_json_finish = build_trace_counter_sm_efficiency(gpu_id, r[0][1], 0)
+                counter_json_str += (efficiency_json_start + efficiency_json_finish)
+        counter_json_bytes = bytes(counter_json_str, 'utf-8')
+        return counter_json_bytes
+
+    def read_file_and_append_gpu_metrics(self, local_file):
+        counter_json_bytes = self.get_gpu_metrics()
+
+        import gzip
+
+        raw_data = io.read(local_file)
+        if local_file.endswith('.gz'):
+            raw_data = gzip.decompress(raw_data)
+        raw_data_without_tail = raw_data[: raw_data.rfind(b']')]
+        raw_data = b''.join([raw_data_without_tail, counter_json_bytes, b']}'])
+
+        raw_data = gzip.compress(raw_data, 1)
+        return raw_data
 
 class DistributedRunProfile(object):
     """ Profiling all workers in a view.
